Refactored comparison main() into a standalone
class called PipelineValidationTool.jar
