+++ src/main/java/util/comparators/CompareReviewDirs.java
@@ -1,24 +1,13 @@
 package util.comparators;
 
-import java.io.File;
-import java.io.FileWriter;
 import java.io.IOException;
-import java.io.PrintStream;
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.HashMap;
 import java.util.LinkedHashMap;
-import java.util.List;
 import java.util.Map;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.logging.Logger;
 
 import json.JSONException;
-import json.JSONObject;
-import net.sourceforge.argparse4j.ArgumentParsers;
-import net.sourceforge.argparse4j.inf.ArgumentParser;
-import net.sourceforge.argparse4j.inf.ArgumentParserException;
-import net.sourceforge.argparse4j.inf.Namespace;
 import operator.IOOperator;
 import pipeline.Pipeline;
 import util.reviewDir.ManifestParseException;
@@ -166,213 +155,6 @@ public class CompareReviewDirs extends IOOperator {
 
 	public void setFinalJSONOutput(LinkedHashMap<String, Object> finalJSONOutput) {
 		this.finalJSONOutput = finalJSONOutput;
-	}
-	
-	/**
-	 * @param args
-	 * @throws ManifestParseException 
-	 * @throws IOException 
-	 * @throws ArgumentParserException 
-	 * @throws JSONException 
-	 */
-	public static void main(String[] args) {
-		ArgumentParser parser = ArgumentParsers.newArgumentParser("CompareReviewDirs.jar")
-			.defaultHelp(true)
-			.description("Perform comparisons of Pipeline review directories, comparing annotations, BAM coverage/metrics, variants, run times, etc. "
-			+ "If given two directories full of RDs this tools performs a comprehensive comparison between all RDs, intelligently comparing runs "
-			+ "that used the same fastq file (so ideally, you have a truth set directories of RDs and a new test set run using a newer"
-			+ "pipeline version and fastq names werent changed). Used in validations, and produces a PASS or FAIL along with a detailed summary.");
-		
-		parser.addArgument("-i", "--input")
-			.nargs(2)
-			.required(true)
-			.help("Enter two paths: Either two RDs or two folders of RDs (in the case of validations).");
-		parser.addArgument("-o", "--out")
-			.nargs(1)
-			.required(true)
-			.help("Name of output file, where a summary of the comparison will be written (will write a json and a summary file).");
-		
-		Namespace ns = null;
-		List<String> paths = null;
-		String outFileName = "";
-		try {
-		    ns = parser.parseArgs(args);
-			paths = ns.getList("input");
-			outFileName = ns.getString("out").replace("[", "").replace("]", "");
-		} catch (ArgumentParserException e) {
-		    parser.handleError(e);
-		    System.exit(1);
-		}
-		System.out.println(paths);
-		System.out.println(outFileName);
-		//check if inputs are valid RDs if not, assume its a validation.
-		try {
-			ReviewDirectory rd1 = new ReviewDirectory(paths.get(0));
-			ReviewDirectory rd2 = new ReviewDirectory(paths.get(1));
-			CompareReviewDirs crd = new CompareReviewDirs(paths.get(0), paths.get(1));
-			crd.compare();
-			
-			try (FileWriter file = new FileWriter(outFileName + ".json")) {
-				file.write(new JSONObject(crd.getFinalJSONOutput()).toString());
-				System.out.println("Successfully Copied JSON Object to File: " + outFileName + ".json");
-			}
-			
-		} catch (ManifestParseException e) {
-			PrintStream out = System.out;
-			if (paths.get(0) == "" || paths.get(1) == "") {
-				out.println("Error with input.");
-				return;
-			}
-			
-			File dir1 = new File(paths.get(0));
-			File dir2 = new File(paths.get(1));
-			if (dir1.isDirectory() && dir2.isDirectory()) {
-				if(dir1.list().length > 0 && dir2.list().length > 0) {
-					// Prepare comparison ------------------------------------------------------------------------------
-					System.out.println("Begining validation of: " + dir1.getName() + " and " + dir2.getName());
-					Map<String, List<ReviewDirectory>> comparisonMap = new HashMap<String, List<ReviewDirectory>>();
-					Map<ReviewDirectory, String> reviewDirPathMap = new HashMap<ReviewDirectory, String>();
-					
-					ArrayList<ReviewDirectory> RDs1 = new ArrayList<ReviewDirectory>();
-					ArrayList<ReviewDirectory> RDs2 = new ArrayList<ReviewDirectory>();
-					
-					for (File f : dir1.listFiles()) {
-						try {
-							ReviewDirectory newRD = new ReviewDirectory(f.getAbsolutePath());
-							RDs1.add(newRD);
-							reviewDirPathMap.put(newRD, f.getAbsolutePath());
-						} catch (IOException | ManifestParseException ex) {
-							ex.printStackTrace();
-						}
-					}
-
-					for (File f : dir2.listFiles()) {
-						try {
-							ReviewDirectory newRD = new ReviewDirectory(f.getAbsolutePath());
-							RDs2.add(newRD);
-							reviewDirPathMap.put(newRD, f.getAbsolutePath());
-						} catch (IOException | ManifestParseException ex) {
-							ex.printStackTrace();
-						}
-					}
-					
-					System.out.println(dir1.getName() + " has " + String.valueOf(RDs1.size()) + " review directories." );
-					System.out.println(dir2.getName() + " has " + String.valueOf(RDs2.size()) + " review directories." );
-					//Now lets populate our comparisonMap.
-					for (ReviewDirectory rd1 : RDs1) {
-						String[] rd1Fastqs = rd1.getLogFile().getFastqNames();
-						
-						for (ReviewDirectory rd2 : RDs2) {
-							String[] rd2Fastqs = rd2.getLogFile().getFastqNames();
-							if( Arrays.equals(rd1Fastqs, rd2Fastqs) ) {
-								List<ReviewDirectory> rds = new ArrayList<ReviewDirectory>();
-								//Make sure the older run RD gets put in the first column as our truth set.
-								if(Long.valueOf(rd1.getSampleManifest().getTime()) < Long.valueOf(rd2.getSampleManifest().getTime())) {
-									rds.add(rd1);
-									rds.add(rd2);
-								} else {
-									rds.add(rd2);
-									rds.add(rd1);
-								}
-								comparisonMap.put(rd1Fastqs[0], rds );
-							}
-						}
-					}
-					// End Prepare comparison --------------------------------------------------------------------------
-					
-					//Start processing summary of comparison -----------------------------------------------------------------
-					Map<String, DiscordanceSummary> valSummary = new HashMap<String, DiscordanceSummary>();
-					
-					//Map<Severity, Integer> severitySummary = new HashMap<Severity, Integer> ();
-					LinkedHashMap<String, Object> validationJSON = new LinkedHashMap<String, Object>();
-
-					for (Map.Entry<String, List<ReviewDirectory>> entry : comparisonMap.entrySet()) {
-						try {
-							CompareReviewDirs crd = new CompareReviewDirs(entry.getValue().get(0).getSourceDirPath(), entry.getValue().get(1).getSourceDirPath());
-							crd.compare();
-							//Now collect relevant summary information from our comparator class.
-							System.out.println("===================================================");
-							String comparisonName = crd.getRd1().getSampleName() + "-" + crd.getRd2().getSampleName();
-							valSummary.put(comparisonName, crd.getDiscordanceSummary());
-							validationJSON.put(comparisonName, crd.getFinalJSONOutput());
-							//valSummary.add(crd.getSummary());
-						} catch (IOException | ManifestParseException | JSONException ex) {
-							System.out.println("Error with comparison for RDs: " + entry.getValue().get(0).getSourceDirPath() + " and " + entry.getValue().get(1).getSourceDirPath());
-							ex.printStackTrace();
-						}
-					}
-					
-					LinkedHashMap<String, Object> validationSummary = new LinkedHashMap<String, Object>();
-					//validationSummary.put("severity.key", getNames(Severity.class));
-					System.out.println("\n\n+++++++++++++++++++++++++");
-					System.out.println("| Summary of Validation |");
-					System.out.println("+++++++++++++++++++++++++");
-
-					for (Severity sev: Severity.values()) {
-						if (!sev.toString().equals("EXACT")) {
-							ComparisonSummaryTable st = new ComparisonSummaryTable();
-							st.setCompareType(sev.toString());
-							st.setColNames(Arrays.asList("#", "Type", ""));
-							LinkedHashMap<String, Object> sevJSON = new LinkedHashMap<String, Object>();
-
-							for (Map.Entry<String, DiscordanceSummary> entry : valSummary.entrySet()) {
-								String comparisonName = entry.getKey();
-								DiscordanceSummary disSum = entry.getValue();
-								
-								List<String> newRow = new ArrayList<>();
-								newRow.add(comparisonName);
-								
-								//newRow.add(sev.toString());
-								Integer sum = 0;
-								for (AtomicInteger i : disSum.getSeveritySummary(sev).values()) {
-								    sum += i.get();
-								}
-								if (sum > 0) {
-									String sevNum = String.valueOf(sum);
-									newRow.add(sevNum);
-									
-									String sevMap = disSum.getSeveritySummary(sev).keySet().toString();
-									newRow.add(sevMap);
-									newRow.add("");
-									
-									String[] summaryArray = {sevNum, sevMap};
-									//validationSummary.put(comparisonName, summaryArray);
-									sevJSON.put(comparisonName, summaryArray);
-									st.addRow(newRow);
-								}				
-							}
-							st.printSeverityTable();
-							validationSummary.put(sev.toString(), sevJSON);
-						}
-					}
-					validationJSON.put("validation", validationSummary);
-					//String jsonString = new JSONObject(validationJSON).toString();
-					//System.out.println(jsonString);
-					
-					// try-with-resources statement based on post comment below :)
-					try (FileWriter file = new FileWriter(outFileName + ".json")) {
-						file.write(new JSONObject(validationJSON).toString());
-						System.out.println("Successfully Copied JSON Object to File: " + outFileName + ".json");
-					} catch (IOException e1) {
-						e1.printStackTrace();
-					}
-					
-					//END processing summary of comparison -----------------------------------------------------------------
-
-				} else {
-					System.out.println("It seems one (or both) the directories given are empty:" + dir1.getName() + " and " + dir2.getName());
-					return;
-				}
-			} else {
-				System.out.println("It seems one (or both) of the inputs are either not directories or don't exist.");
-				return;
-			}
-			
-		} catch (IOException e) {
-			e.printStackTrace();
-		} catch (JSONException e) {
-			e.printStackTrace();
-		}
 	}
 	
 	@Override
